cookiecutter-rag/{{cookiecutter.project_slug}}/pyproject.toml
[project]
name = "{{ cookiecutter.project_slug }}"
version = "{{ cookiecutter.version }}"
description = "{{ cookiecutter.project_description }}"
authors = [
    {name = "{{ cookiecutter.author_name }}", email = "{{ cookiecutter.author_email }}"},
]
requires-python = "{{ cookiecutter.python_version }}"
dependencies = [
    "kitchenai-whisk",
    "kitchenai-llama",
    "llama-index",
    "llama-index-vector-stores-chroma",
    "chromadb",
    "tiktoken",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "isort",
    "mypy"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"] 